Fixed attributes scan; moved all dataclasses to one place
